+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [#184](https://github.com/wandersoncferreira/code-review/pull/184): [gitlab] remove requirement for feedback to send only comments
 - [4ccad8c](https://github.com/wandersoncferreira/code-review/commit/4ccad8cb89ad38075a23023affbac1eb545a99ab): [gitlab] send feedback comment. Sorry commit on master, too tired :/
 - [#185](https://github.com/wandersoncferreira/code-review/pull/185): [gitlab] set title
+- [#187](https://github.com/wandersoncferreira/code-review/pull/187): [github] feature add - suggestion box bound to `C-c C-s`
 
 # v0.0.6